Drop per package CPU/GPU mutexes
Based on discussions with the xgboost developers, it appears that only
the C++ library changes based on whether it is built in CPU or GPU mode.
The other language bindings do not change based on whether this
functionality exists or not. So drop all of the per package mutexes to
streamline things a bit.