Add GHA workflow stub for uploading packages
We would like to use the `workflow_run` event to add a custom conda
package upload workflow after packages are built. However in the GitHub
Actions (GHA) docs, it states that `workflow_run` events will only be
triggered if the workflow is on the default branch. Though we have not
figured out completely how to implement this workflow and don't want to
add it untested to `main`. So simply add a stub workflow with the
`workflow_run` event. This way GHA will run the workflow, but we can
start customizing it in a PR instead where we can actually test its
behavior.

https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_run